+++ src/welcome/WelcomeImages.js
@@ -27,8 +27,8 @@ export const GetInvolvedImage = props => (
 
 const styles = StyleSheet.create({
   candidates: {
-    height: 180,
-    width: 180,
+    height: 138,
+    width: 138,
   },
   stayInformed: {
     height: 148,
